test_basic_signature.py
#!/usr/bin/env python3
"""
Test simplifié pour vérifier notre implémentation SignedDocumentCreator.
"""

import sys
import os
sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'app'))

def test_basic_functionality():
    """Test les fonctionnalités de base sans dépendances complexes."""
    print("🔍 Test des fonctionnalités de base...")
    
    try:
        # Test borb de base
        from borb.pdf import PDF, Document
        print("✅ borb PDF et Document importés")
        
        # Test cryptography
        from cryptography.hazmat.primitives import hashes
        from cryptography.hazmat.primitives.asymmetric import rsa, padding
        from cryptography.hazmat.primitives.serialization import Encoding, PrivateFormat, NoEncryption
        print("✅ cryptography importé")
        
        # Test PIL
        from PIL import Image as PILImage
        print("✅ Pillow importé")
        
        return True
        
    except Exception as e:
        print(f"❌ Erreur : {e}")
        return False

def test_certificate_creation_only():
    """Test uniquement la création de certificats RSA."""
    print("\n🔐 Test de création de certificats RSA...")
    
    try:
        from cryptography.hazmat.primitives.asymmetric import rsa
        from cryptography.hazmat.primitives import hashes, serialization
        from cryptography.hazmat.primitives.asymmetric import padding
        import hashlib
        import datetime
        import secrets
        
        # Générer une clé RSA de test
        print("   Génération d'une clé RSA...")
        private_key = rsa.generate_private_key(
            public_exponent=65537,
            key_size=2048
        )
        public_key = private_key.public_key()
        
        # Créer un hash de test
        test_data = "Test document content"
        document_hash = hashlib.sha256(test_data.encode()).hexdigest()
        
        # Signer le hash
        print("   Signature du document...")
        signature = private_key.sign(
            document_hash.encode(),
            padding.PSS(
                mgf=padding.MGF1(hashes.SHA256()),
                salt_length=padding.PSS.MAX_LENGTH
            ),
            hashes.SHA256()
        )
        
        # Vérifier la signature
        print("   Vérification de la signature...")
        try:
            public_key.verify(
                signature,
                document_hash.encode(),
                padding.PSS(
                    mgf=padding.MGF1(hashes.SHA256()),
                    salt_length=padding.PSS.MAX_LENGTH
                ),
                hashes.SHA256()
            )
            print("✅ Signature RSA vérifiée avec succès")
            return True
        except Exception as e:
            print(f"❌ Échec de la vérification : {e}")
            return False
            
    except Exception as e:
        print(f"❌ Erreur lors du test RSA : {e}")
        return False

def test_svg_simple():
    """Test SVG sans borb."""
    print("\n🎨 Test SVG basique...")
    
    try:
        # Test cairosvg seulement
        import io
        from PIL import Image as PILImage
        
        # SVG simple
        test_svg = '''<svg width="100" height="50" xmlns="http://www.w3.org/2000/svg">
            <text x="10" y="30" fill="blue" font-family="Arial" font-size="14">Signature Test</text>
        </svg>'''
        
        # Pour l'instant, juste créer une image vide pour simuler
        print("   Création d'une image de test...")
        test_image = PILImage.new('RGBA', (100, 50), (255, 255, 255, 0))
        
        print("✅ Test SVG basique réussi")
        print(f"   Taille de l'image : {test_image.size}")
        return True
        
    except Exception as e:
        print(f"❌ Erreur SVG : {e}")
        return False

def main():
    """Fonction principale de test."""
    print("=== Test simplifié de SignedDocumentCreator ===\n")
    
    tests_results = []
    
    # Test 1: Imports de base
    print("1. Test des imports de base :")
    tests_results.append(test_basic_functionality())
    
    # Test 2: Certificats RSA
    print("\n2. Test des certificats RSA :")
    tests_results.append(test_certificate_creation_only())
    
    # Test 3: SVG basique
    print("\n3. Test SVG basique :")
    tests_results.append(test_svg_simple())
    
    # Résumé
    print("\n" + "="*50)
    passed = sum(tests_results)
    total = len(tests_results)
    
    if passed == total:
        print(f"🎉 Tous les tests de base sont passés ! ({passed}/{total})")
        print("\n✨ Les composants fondamentaux fonctionnent :")
        print("   • borb pour PDF")
        print("   • cryptography pour RSA")
        print("   • PIL pour les images")
    else:
        print(f"⚠️  {passed}/{total} tests passés")
        
    return passed == total

if __name__ == "__main__":
    success = main()
    if success:
        print("\n🚀 Prêt à tester la classe complète SignedDocumentCreator !")
    else:
        print("\n🛠️  Dépendances à corriger avant de continuer.")