requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dotenv==0.9.9
Flask==3.1.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
paramiko==3.5.1
pycparser==2.22
pycups==2.0.4
PyNaCl==1.5.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
pandas==2.3.2
alembic==1.16.5
waitress==3.0.2
borb==3.0.2
pillow==11.3.0
cairosvg==2.8.2